Introduction

A small package allows to index arrays by a function via a simple wrapper FI. Besides, for indexing with !=(n), !in(itr), there is another wrapper not providing a convenient and optimized way, which is similar to Not of InvertedIndices., but faster in some cases.

References

FunctionIndices.AbstractFunctionIndex — Type

AbstractFunctionIndex

Supertype of all function index types.

FunctionIndices.AbstractNotIndex — Type

AbstractNotIndex{T}

Supertype of all not types which create by not.

FunctionIndices.FunctionIndex — Type

FunctionIndex{F} <: AbstractFunctionIndex
FI

A implementation of function index, which make indexing with function possible. 1-d indexing A[FI(f)] is equivalent to A[map(f, begin:end)], multi-dimensional indexing A[FI(f1), ..., FI(fn)] is equivalent to A[map(FI(f1), axes(A, 1)), ..., map(FI(fn), axes(A, n))].

FunctionIndices.NotIndex — Type

NotIndex{T} <: AbstractNotIndex{T}

The default implementation of not. There are some optimization for NotIndex(x) comparing to FI(!in(x)). For large arrays, this implementation may be faster. but for small arrays this implementation may be slower. See performance comparing for more details.

Optimization can be enable/disable by override optimized.

FunctionIndices.not — Method

not(x)

Create a NotIndex with given x, which is similar to Not of InvertedIndices. In most cases, not is much faster than Not. See performance comparing for more details.

Main differences between not and Not

For CartesianIndex, A[not(CartesianIndex(i, j,..., n))] is equivalent to A[not(i), not(j), ..., not(n)] and will return a array with same dimension, but A[Not(CartesianIndex(i, j,..., n))] will convert the CartesianIndex to a linear index and return a vector:

julia> A = reshape(1:12, 3, 4)
3×4 reshape(::UnitRange{Int64}, 3, 4) with eltype Int64:
 0  3  6   9
 1  4  7  10
 2  5  8  11

julia> A[not(CartesianIndex(1, 2))] # equivalent to A[not(1), not(2)]
2×3 Matrix{Int64}:
 1  7  10
 2  8  11

julia> A[Not(CartesianIndex(1, 2))] # equivalent to A[Not(3)]
11-element Vector{Int64}:
  0
  1
  2
  4
  5
  ⋮
  8
  9
 10
 11

Besides, for index like A[4, 5] which is out of bounds, A[not(4), not(5)] is equivalent to A[:, :], because inbounds indices are always not equal to the given value, while A[Not[4], Not(5)] causes an error.

FunctionIndices.optimized — Method

optimized(A, I) -> StaticBool
optimized(::Type{TA}, ::Type{TI}) -> StaticBool

Determine whether try optimized indexing for given type of TA and TI. By default, it's False() for TI<:AbstractFunctionIndex and True() for TI<:NotIndex.

Override optimized(::Type{TA}, ::Type{TI}) to enable/disable optimized indexing for given type.

FunctionIndices.to_index — Method

FunctionIndices.to_index(::StaticBool, ind, i)

Convert a AbstractFunctionIndex i to a array index with ind. The first argument of an optimized methods should be True, like to_index(::True, ind, i).