Introduction

A small package allows to access array elements by a function via a simple wrapper FI. As a special case, for indexing with !=(n), !in(itr), there is another wrapper not providing a convenient and optimized way. The not is similar to Not of InvertedIndices, but faster in some cases. Besides, this package also provides ways to change the behavior for special array types and function index types.

Quick start

1-d indexing A[FI(f)] is equivalent to A[map(f, begin:end)], multi-dimensional indexing A[FI(f1), ..., FI(fn)] is equivalent to A[map(FI(f1), axes(A, 1)), ..., map(FI(fn), axes(A, n))].

julia> using FunctionIndices
julia> A = reshape(0:11, 3, 4)3×4 reshape(::UnitRange{Int64}, 3, 4) with eltype Int64:
 0  3  6   9
 1  4  7  10
 2  5  8  11
julia> A[FI(iseven)]6-element Vector{Int64}:
  1
  3
  5
  7
  9
 11
julia> A[map(iseven, begin:end)]6-element Vector{Int64}:
  1
  3
  5
  7
  9
 11
julia> A[FI(isodd), FI(iseven)]2×2 Matrix{Int64}:
 3   9
 5  11
julia> A[map(isodd, begin:end), map(iseven, begin:end)]2×2 Matrix{Int64}:
 3   9
 5  11

not is an alternative to Not and in most cases they are equivalent:

julia> using InvertedIndices
julia> A[not(1)] == A[Not(1)]true
julia> A[not(1, 2)] == A[Not(1, 2)]true
julia> A[not(1:2)] == A[Not(1:2)]true
julia> let bools = rand(Bool, size(A)); A[not(bools)] == A[Not(bools)] endtrue

But for CartesianIndex, A[not(CartesianIndex(i, j,..., n))] is equivalent to A[not(i), not(j), ..., not(n)] and will return a array with same dimension but A[Not(CartesianIndex(i, j,..., n))] will convert the CartesianIndex to a linear index and return a vector:

julia> A[not(CartesianIndex(1, 2))] # equivalent to A[not(1), not(2)]2×3 Matrix{Int64}:
 1  7  10
 2  8  11
julia> A[Not(CartesianIndex(1, 2))] # equivalent to A[Not(3)]11-element Vector{Int64}:
  0
  1
  2
  4
  5
  6
  7
  8
  9
 10
 11

Besides, for out of bounds index like A[4, 5], A[not(4), not(5)] is equivalent to A[:, :], because inbounds indices are not equal to the given value, while A[Not[4], Not(5)] causes an error.

Performant tips for not

For small array, the optimized not(x) might be slower in some case, see performance comparing for details.

There are some tips for better performance:

Use FI(!in(x)) instead of not(x).

Create your own "Not" type, see below example for details.

For a small array of indices like not([1, 2, 3]), not(1, 2, 3) will faster.

Mechanism

This package define a type AbstractFunctionIndex (AFI for short) which can be convert ed to a array index by Base.to_indices.

There are three methods determining how to convert AFI to array index:

FunctionIndices.to_index: this function is the direct method converting AFI. The default method of to_index converting AFI to a function and map it for ind. Besides, the to_index also accepts a Type argument, which is determining call which method of to_index to convert AFI.

FunctionIndices.to_function: this function is called in default method to_index and convert the given AFI to a function.

FunctionIndices.indextype: this function is called in to_indices and returns a type as the Type argument of to_index. The indextype accepts two arguments, the type of array and type of a AFI.

Example to define "Not"

If you don't like the default behavior of not, creating a new "Not" index type is easy:

using FunctionIndices
struct YatAnotherNotIndex{T} <: FunctionIndices.AbstractNotIndex{T}
    parent::T
end
const YANI = YatAnotherNotIndex
Base.parent(I::YatAnotherNotIndex) = I.parent
reshape(1:10, 2, 5)[YANI(1), YANI(2)]

1×4 Matrix{Int64}:
 2  6  8  10

Info

to_function for AbstractNotIndex is defined as !in(parent(I)).

If a big array of linear indices I should be exclude, create a new index array by setdiff might faster than map(!in(I)). You can do this by Defining indextype and to_index for YANI:

FunctionIndices.indextype(::Type{<:AbstractArray}, ::Type{<:YANI{<:Array{<:Integer}}}) = Vector{Int}
FunctionIndices.to_index(::Type{Vector{Int}}, ind, I::YANI{<:Array{<:Integer}}) = setdiff(ind, parent(I))::Vector{Int}
typeof(to_indices(0:10, (YANI([1, 2, 3]),))[1]), typeof(to_indices(0:10, (not([1, 2, 3]),))[1])

(Vector{Int64}, Base.LogicalIndex{Int64, MappedArrays.ReadonlyMappedArray{Bool, 1, Base.OneTo{Int64}, Base.var"#76#77"{Base.Fix2{typeof(in), Set{Int64}}}}})

References

FunctionIndices.AbstractFunctionIndex — Type

AbstractFunctionIndex

Supertype of all function index types.

FunctionIndices.AbstractNotIndex — Type

AbstractNotIndex{T}

Supertype of all not types which create by not.

FunctionIndices.FunctionIndex — Type

FunctionIndex{F} <: AbstractFunctionIndex
FI

A implementation of function index, which make indexing with function possible. 1-d indexing A[FI(f)] is equivalent to A[map(f, begin:end)], multi-dimensional indexing A[FI(f1), ..., FI(fn)] is equivalent to A[map(FI(f1), axes(A, 1)), ..., map(FI(fn), axes(A, n))].

FunctionIndices.NotIndex — Type

NotIndex{T} <: AbstractNotIndex{T}

The default implementation of not. There are some optimization for NotIndex(x) comparing to FI(!in(x)). For large arrays, this implementation may be faster. but for small arrays this implementation may be slower. See performance comparing for more details.

FunctionIndices.indextype — Method

indextype([A::AbstractArray,] I::AbstractFunctionIndex)
indextype([::Type{TA},] ::Type{TI})

Determine the index type which the function index type TI will be converted to by Base.to_indices for array type TA. By default, it's AbstractArray.

Note

If you define a methods indextype(::Type{TA}, ::Type{TI}) = T, while to_index(::Type{T}, inds, I::TI) is not defined, the to_indices will not convert a index of type TI to a T, but a Base.LogicalIndex, the default return type of to_index.

FunctionIndices.not — Method

not(x)

Create a NotIndex with given x, which is similar to Not of InvertedIndices. In most cases, not is much faster than Not. See performance comparing for more details.

Main differences between not and Not

For CartesianIndex, A[not(CartesianIndex(i, j,..., n))] is equivalent to A[not(i), not(j), ..., not(n)] and will return a array with same dimension, but A[Not(CartesianIndex(i, j,..., n))] will convert the CartesianIndex to a linear index and return a vector:

julia> A = reshape(1:12, 3, 4)
3×4 reshape(::UnitRange{Int64}, 3, 4) with eltype Int64:
 0  3  6   9
 1  4  7  10
 2  5  8  11

julia> A[not(CartesianIndex(1, 2))] # equivalent to A[not(1), not(2)]
2×3 Matrix{Int64}:
 1  7  10
 2  8  11

julia> A[Not(CartesianIndex(1, 2))] # equivalent to A[Not(3)]
11-element Vector{Int64}:
  0
  1
  2
  4
  5
  ⋮
  8
  9
 10
 11

Besides, for index like A[4, 5] which is out of bounds, A[not(4), not(5)] is equivalent to A[:, :], because inbounds indices are always not equal to the given value, while A[Not[4], Not(5)] causes an error.

FunctionIndices.to_function — Function

to_function(I::AbstractFunctionIndex)

Converts a AbstractFunctionIndex to a Function. By default, to_function(I::AbstractNotIndex) returns !in(parent(I)).

FunctionIndices.to_index — Method

FunctionIndices.to_index(::Type{T<:AbstractArray}, ind, i)

Convert a AbstractFunctionIndex i to a array index of type T with ind. By default, to_index(::AbstractArray, ind, i) will return a Base.LogicalIndex{Bool, ReadonlyMappedArray{Bool...}}.